Use double in ex 2.2

+++ 2_les_briques_de_base_d_un_programme/2.2_le_rectangle_dort.c
@@ -3,12 +3,12 @@
 #define PI 3.14159265359
 
 int main(void) {
-    float rayon;
+    double rayon;
 
     printf("Quelle est la taille du rayon du cercle dont vous voulez connaitre la circonférence ? ");
-    scanf("%f", &rayon);
+    scanf("%lf", &rayon);
 
-    printf("La circonférence du cercle vaut: %f\n", 2 * rayon * PI);
+    printf("La circonférence du cercle vaut: %lf\n", 2 * rayon * PI);
 
     return 0;
 }
